// src/app/models/master-subject-model.ts
export class MasterSubjectModel {
    masterSubjectId: number;
    schoolId: number;
    subjectName: string;
    subjectCode: string;
    marksType: number;
    subjectType: number;
    displayOrder: number;
    activeStatus: number;
    createdByUserId: number;
    updatedByUserId: number;
    createdDate:Date;        
    updatedDate:Date; 

}